refactor: adjust public api
